#79 - py.test and whitenoise check
Another problem came up regarding the STAGE=test flag, so that py.test runs with Whitenoise.

py.test environments are not loaded before django settings setup (check https://github.com/MobileDynasty/pytest-env/issues/3). The pytest-env needs to be fixed but a PR is not accepted, I forked the repository and used that one.